% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_transcriptome_loom.R
\name{gmm_cell_caller}
\alias{gmm_cell_caller}
\title{Function for calling cells using gaussian mixture models}
\usage{
gmm_cell_caller(counts, k = 1:6, log_convert = TRUE, verbose = TRUE)
}
\arguments{
\item{counts}{When 'counts' is an atomic integer vector, the values should
represent read counts per cell (e.g. cDNA or gDNA counts). A matrix or
dataframe can also be given, in which case the rows correspond with cells
and the columns represent separate modalities (e.g. RNA and DamID). Thus,
cell class membership can be determined based on multiple modalities.}

\item{k}{An atomic vector of type integer. A gmm is estimated for each
value of 'k', where this value represents the number of components in
the gmm. The cell classifications will be based on the model with the lowest
BIC. To prevent testing of multiple components and to select this yourself,
provide a single integer. Default 'k' is '1:6'.}

\item{log_convert}{Logical value indicating if the counts should be
log-transformed before model estimation.}

\item{verbose}{Logical value indicating if the function should message the
chosen number of components and the corresponding bic.}
}
\value{
A two-element list. The first element is an atomic integer vector,
where the integers represent class membership to one of the predicted
classes (e.g. 1: 'no cell', 2: 'real cell'). The second element represents
the 'Mclust' object that was used for the class membership prediction.
}
\description{
Function for calling cells using gaussian mixture models
}
