% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_transcriptome_loom.R
\name{read_loom}
\alias{read_loom}
\title{Connect to loom file}
\usage{
read_loom(loom_path)
}
\arguments{
\item{loom_path}{A single-element character vector that indicates the path
to the transcriptome loom file.}
}
\value{
An object of class 'loom' that represents a connection to the loom
file stored on disk.
}
\description{
Connect to loom file
}
\examples{
lpath <- "Path/to/loom"

lfile <- read_loom(lpath)
}
