% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_transcriptome_loom.R
\name{MultiLoomAsSeurat}
\alias{MultiLoomAsSeurat}
\title{Convert multiple loom files to a single merged Seuratobject}
\usage{
MultiLoomAsSeurat(
  loom_paths,
  matrix_rowname_col = "Gene",
  matrix_colname_col = "CellID",
  seurat_assay_name = "RNA",
  seurat_min_cells = 0,
  seurat_names_field = 1L,
  seurat_names_delim = "_",
  ...
)
}
\arguments{
\item{loom_paths}{A character vector that indicates the paths to the
transcriptome loom files.}

\item{matrix_rowname_col}{A single element character vector that indicates
the variable (i.e. column) of the row (i.e. gene) metadata in the loom file.
This variable will subsequently be used as the feature names for the
generated sparse matrix.}

\item{matrix_colname_col}{A single element character vector that indicates
the variable (i.e. column) of the column (i.e. cell) metadata in the loom
file. This variable will subsequently be used as the cell names for the
generated sparse matrix.}

\item{seurat_assay_name}{A single element character vector representing the
name of the assay in the Seuratobject.}

\item{seurat_min_cells}{Include only features in the seuratobject that are
detected in at least this many cells.}

\item{seurat_names_field}{Argument used to get sample name from the cell
names. Works in conjunction with the 'seurat_names_delim' argument.
'seurat_names_field' is an integer selecting which substring
(i.e. delimited by 'seurat_names_delim') should be selected as the sample
name. The default is 1L (i.e. the first field of the cell name).}

\item{seurat_names_delim}{Argument used to get sample name from the cell
names. Works in conjunction with the 'seurat_names_field' argument.
'seurat_names_delim' is a single element character vector that represents
the delimiter used to select the cell name substrings.}

\item{...}{Additional arguments will be given to the
SeuratObject::CreateSeuratObject() function call.}
}
\value{
A Seuratobject populated with a sparse matrix representing the
transcriptome data from the loom files located at 'loom_paths'.
The 'orig.idents' variable in the cell metadata is used to identify from
which sample each cell came from in the merged object. This 'orig.idents'
gets chosen based on the 'seurat_names_field' and 'seurat_names_delim'
arguments.
}
\description{
Convert multiple loom files to a single merged Seuratobject
}
\examples{
lpaths <- c("Path/to/loomfile1",
            "Path/to/loomfile2",
            "Path/to/loomfile3")

lseurat <- LoomAsSeurat(lpath,
  matrix_rowname_col = "Gene",
  matrix_colname_col = "CellID",
  seurat_assay_name = "RNA",
  seurat_min_cells = 0,
  seurat_names_field = 1L,
  seurat_names_delim = "_")
}
